requirements.txt
streamlit
openai
pandas
python-dotenv
requests
supabase
Pillow
